+++ orttraining/orttraining/training_ops/cuda/math/isfinite.h
@@ -2,9 +2,9 @@
 // Licensed under the MIT License.
 
 #pragma once
+
 #include "core/common/common.h"
 #include "core/providers/cuda/cuda_kernel.h"
-#include "core/providers/cuda/multi_tensor/common.cuh"
 
 namespace onnxruntime {
 namespace cuda {
@@ -17,9 +17,6 @@ class IsFiniteOp final : public CudaKernel {
 
   Status ComputeInternal(OpKernelContext* context) const override;
 };
-
-template <typename TSrc>
-void IsFinite(cudaStream_t stream, const TSrc* input, bool* output, size_t N);
 
 }  // namespace cuda
 }  // namespace onnxruntime
