+++ orttraining/orttraining/training_ops/cuda/math/scale.cc
@@ -2,6 +2,7 @@
 // Licensed under the MIT License.
 
 #include "orttraining/training_ops/cuda/math/scale.h"
+#include "orttraining/training_ops/cuda/math/scale_impl.h"
 
 using namespace ONNX_NAMESPACE;
 using namespace onnxruntime::common;
@@ -24,6 +25,15 @@ namespace cuda {
                                      DataTypeImpl::GetTensorType<int32_t>()})  \
           .InputMemoryType(OrtMemTypeCPUInput, 1),                             \
       Scale<T>);
+
+template <typename ScaleT>
+struct GetScaleValueImpl {
+  void operator()(const Tensor* scale, float& scale_value) const {
+    ORT_ENFORCE(scale->Shape().Size() == 1, "Scale input should have a single value.");
+    scale_value = static_cast<float>(*(scale->template Data<ScaleT>()));
+    ORT_ENFORCE(scale_value != 0.0f, "Scale value must not be 0.");
+  }
+};
 
 template <typename T>
 Scale<T>::Scale(const OpKernelInfo& info) : CudaKernel(info) {
