+++ orttraining/orttraining/training_ops/cuda/reduction/reduction_all.h
@@ -3,7 +3,6 @@
 
 #pragma once
 #include "core/providers/cuda/cuda_kernel.h"
-#include "core/providers/cuda/multi_tensor/common.cuh"
 
 namespace onnxruntime {
 namespace cuda {
@@ -15,14 +14,6 @@ class ReduceAllL2 final : public CudaKernel {
 
   Status ComputeInternal(OpKernelContext* context) const override;
 };
-
-template <typename TIn, typename TOut>
-struct MultiTensorReduceL2 {
-  void operator()(cudaStream_t stream, ChunkGroup<1> chunk_group, TOut* output);
-};
-
-template<typename Tin, typename Tout>
-void ScalarSqrt(cudaStream_t stream, Tin* input, Tout* output);
 
 }  // namespace cuda
 }  // namespace onnxruntime
