+++ orttraining/orttraining/training_ops/rocm/math/softmax_grad.cc
@@ -2,6 +2,7 @@
 // Licensed under the MIT License.
 
 #include "orttraining/training_ops/rocm/math/softmax_grad.h"
+#include "orttraining/training_ops/rocm/math/softmax_grad_impl.h"
 
 #include "core/providers/common.h"
 #include "core/providers/rocm/miopen_common.h"
