+++ orttraining/orttraining/training_ops/rocm/math/softmax_grad_impl.cu
@@ -19,7 +19,7 @@
 // The code below is mostly copied from Pytorch PersistentSoftmax.cuh
 #include "hip/hip_runtime.h"
 
-#include "orttraining/training_ops/rocm/math/softmax_grad.h"
+#include "orttraining/training_ops/rocm/math/softmax_grad_impl.h"
 
 #include "core/providers/rocm/cu_inc/common.cuh"
 #include "core/providers/rocm/math/softmax_warpwise_impl.cuh"
